feat: webhook 임의로 구현

+++ src/main/java/io/github/tlsdla1235/seniormealplan/domain/Meal.java
@@ -25,7 +25,7 @@ public class Meal {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "meal_id")
-    private Integer mealId;
+    private Long mealId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
